+++ domain/src/main/java/com/example/domain/data/CheckAuthRequestResult.kt
@@ -7,6 +7,6 @@ data class CheckAuthRequestResult(
     val syntaxErrorDetails: List<SyntaxErrorDetail>? = listOf(),
     val refreshToken: String = "",
     val accessToken: String = "",
-    val userId: Long = 0L,
+    val userId: Int = 0,
     val isUserExists: Boolean = false,
 )
