+++ domain/src/main/java/com/example/domain/data/CheckAuthResult.kt
@@ -4,7 +4,7 @@ data class CheckAuthResult(
     val success: Boolean = false,
     val refreshToken: String = "",
     val accessToken: String = "",
-    val userId: Long = 0L,
+    val userId: Int = 0,
     val isUserExists: Boolean = false,
     val errorText: String = "",
 )
